fix: promise callback queue and example api for preview server
Adding promise callback handling to bind to the existing promise if there is already a promise in the works.

This prevents the callback from being lost when there is already a request in process.

Also fixed the example api to use a 'preview server' for the `/content/pages/index.yaml` in the example to simulate the process of using the api to load the preview server settings to determine the serving url of a page.