feat: Validation delay for forms and fields. (#223)
For a better user experience, wait to validate form fields until the user 'leaves' the field. This provides a more immediate feedback for the user on the value without distracting the user during input.

Also enables the form 'submit' and then marking all errors before disabling the button if invalid.